+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- ui: v-align label in EditableInput
+- ui: refactor EditableSelect into DropdownMenu in AppTimeSettingsModal
+### Added
+- ui: implement calendar header/body/text color pickers in AppTimeSettingsModal
 
 ## [0.1.2] - 2023-10-16
 ### Changed