+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ui: implement weekday active/inactive color pickers in AppTimeSettingsModal
 - ui: implement uppercase toggle in DisplayCard
+- ui: implement "start week on Monday" toggle in AppTimeSettingsModal
 
 ## [0.1.3] - 2023-10-27
 ### Added