+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.4] - 2023-10-28
 ### Added
 - ui: implement weekday active/inactive color pickers in AppTimeSettingsModal
 - ui: implement uppercase toggle in DisplayCard
@@ -47,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/awtrix-light/hub/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/awtrix-light/hub/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/awtrix-light/hub/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/awtrix-light/hub/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/awtrix-light/hub/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/awtrix-light/hub/compare/v0.1.0...v0.1.1